MAUI Blazor - Add Quiz component, update view models and move them to Shared folder

+++ src/FlashBeat.Maui/Components/Shared/DeezePlayerComponent.razor.cs
@@ -9,7 +9,7 @@ public sealed partial class DeezePlayerComponent
 {
     [Parameter]
     [EditorRequired]
-    public int TrackId { get; set; }
+    public long TrackId { get; set; }
 
     /// <summary>
     /// Gets or sets the size (height and width), in px, of the component.
